// app/components/types.ts
export type SegmentType = {
  id: string;
  name: string;
  durationType: string;
  duration: string;
  durationUnit: string;
  intensityType: string;
  intensity: string;
};
export type FormDataType = {
  name: string;
  durationType: string;
  duration: string;
  durationUnit: string;
  intensityType: string;
  intensity: string;
};